Bump build number to 1
This is done as a precaution given the checksum of the GitHub archive
has changed. So we want to indicate this is a build of the package from
a potentially different source. Did compare the source to the git
repository and it does seem to checkout though. So there probably isn't
anything to worry about.